requirements.txt
pyspark==3.5.0
pendulum~=2.1.2
requests~=2.27.1
apache-airflow~=2.5.1
cffi~=1.15.0
delta-spark==3.0.0
pytest==7.1.2
chispa==0.9.2
parameterized==0.8.1
pytest-subtests==0.9.0
pytest-timeout==2.1.0
confluent-kafka==1.9.2
pydantic==1.10.2
python-dateutil==2.8.2
sentry-sdk==1.19.0
tenacity==8.0.1
pandas==1.5.3
numpy==1.23.5
scipy==1.10.1
scikit-learn==1.3.1
pyarrow==12.0.1